ref(TPC) Move all the utility functions to TPCUtils.
